Rename models to fix issue with YOLOv5 from torch hub
