+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "models/detectors/yolo"]
-	path = models/detectors/yolo
+	path = models_/detectors/yolo
 	url = https://github.com/eriklindernoren/PyTorch-YOLOv3
