+++ models_/detectors/YOLOv3.py
@@ -7,7 +7,7 @@ import numpy as np
 import torch
 from torchvision.transforms import transforms
 
-sys.path.append(os.path.join(os.getcwd(), 'models', 'detectors', 'yolo'))
+sys.path.append(os.path.join(os.getcwd(), 'models_', 'detectors', 'yolo'))
 
 from .yolo.models import Darknet
 from .yolo.utils.utils import load_classes, non_max_suppression
